Split the Wsv functions
