+++ src/CMakeLists.txt
@@ -2,27 +2,26 @@ add_subdirectory(partfun)
 add_subdirectory(python_interface)
 add_subdirectory(core)
 
-if (NETCDF_FOUND)
-  set (NETCDF_CC_FILES
+if(NETCDF_FOUND)
+  set(NETCDF_CC_FILES
     nc_io.cc
     nc_io_array_types.cc
     nc_io_basic_types.cc
     nc_io_compound_types.cc
-    )
-endif (NETCDF_FOUND)
+  )
+endif(NETCDF_FOUND)
 
-if (ENABLE_MPI)
-  include_directories (${MPI_CXX_INCLUDE_PATH})
-endif (ENABLE_MPI)
+if(ENABLE_MPI)
+  include_directories(${MPI_CXX_INCLUDE_PATH})
+endif(ENABLE_MPI)
 
-file (GLOB HEADERFILES "${CMAKE_CURRENT_SOURCE_DIR}/*.h")
+file(GLOB HEADERFILES "${CMAKE_CURRENT_SOURCE_DIR}/*.h")
 add_custom_target(UtilityHeadersArts SOURCES ${HEADERFILES})
 
-########### target to extract version number from ChangeLog ###############
+# ########## target to extract version number from ChangeLog ###############
+add_custom_target(auto_version_cc DEPENDS UpdateAutoVersion)
 
-add_custom_target (auto_version_cc DEPENDS UpdateAutoVersion)
-
-add_custom_command (
+add_custom_command(
   OUTPUT auto_version.cc
   COMMAND ${CMAKE_COMMAND} -D "ARTS_BINARY_DIR:STRING=\"${ARTS_BINARY_DIR}\"" -D "ARTS_SOURCE_DIR:STRING=\"${ARTS_SOURCE_DIR}\"" -P ${ARTS_SOURCE_DIR}/cmake/scripts/update_auto_version_cc.cmake
   BYPRODUCTS auto_version_cc.txt
@@ -31,8 +30,7 @@ add_custom_command (
   COMMENT "Updating ARTS version number"
 )
 
-#####
-
+# ####
 add_library(workspace_group STATIC
   workspace_groups.cpp
 )
@@ -46,8 +44,7 @@ add_custom_command(OUTPUT auto_wsg.h auto_wsg.cpp auto_workspace.cpp
 )
 target_link_libraries(make_auto_wsg PRIVATE workspace_group)
 
-#####
-
+# ####
 add_library(workspace_internals STATIC
   workspace_agenda_class_default_constructor.cpp
   workspace_agendas.cpp
@@ -55,14 +52,21 @@ add_library(workspace_internals STATIC
   workspace_methods.cpp
   workspace_variables.cpp
   workspace_variables_keywords.cpp
+  workspace_wsv_ctor.cpp
+  workspace_wsv_dtor.cpp
+  workspace_wsv_copy_ctor.cpp
+  workspace_wsv_move_ctor.cpp
+  workspace_wsv_copy_set.cpp
+  workspace_wsv_move_set.cpp
+  workspace_wsv_copy_fn.cpp
+  workspace_wsv_other_fn.cpp
   auto_wsg.cpp
 )
 target_link_libraries(workspace_internals PUBLIC workspace_group artscore)
 target_include_directories(workspace_internals PUBLIC ${CMAKE_CURRENT_BINARY_DIR})
 target_include_directories(workspace_internals PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
 
-#####
-
+# ####
 add_executable(make_auto_wsv make_auto_wsv.cpp)
 add_custom_command(OUTPUT auto_wsv.h auto_wsv.cpp
   COMMAND make_auto_wsv
@@ -72,8 +76,7 @@ add_custom_command(OUTPUT auto_wsv.h auto_wsv.cpp
 )
 target_link_libraries(make_auto_wsv PRIVATE workspace_internals)
 
-#####
-
+# ####
 if(NOT NUM_ARTS_WSM)
   set(NUM_ARTS_WSM 10)
 endif()
@@ -97,8 +100,7 @@ add_custom_command(OUTPUT auto_wsm.h auto_wsm.cpp auto_wsmmeta.cpp ${ARTS_WORKSP
 )
 target_link_libraries(make_auto_wsm PRIVATE workspace_internals)
 
-#####
-
+# ####
 add_executable(make_auto_wsa make_auto_wsa.cpp)
 add_custom_command(OUTPUT auto_wsa.h auto_wsa.cpp
   COMMAND make_auto_wsa
@@ -108,9 +110,8 @@ add_custom_command(OUTPUT auto_wsa.h auto_wsa.cpp
 )
 target_link_libraries(make_auto_wsa PRIVATE workspace_internals)
 
-########### next target ###############
-
-add_library (artsworkspace STATIC
+# ########## next target ###############
+add_library(artsworkspace STATIC
   auto_wsv.cpp
   auto_wsm.cpp
   auto_wsmmeta.cpp
@@ -208,53 +209,53 @@ add_library (artsworkspace STATIC
   xml_io_rtepack.cc
   xml_operators.cc
   ${NETCDF_CC_FILES}
-        m_fluxes.cc)
+  m_fluxes.cc)
 
-target_include_directories (artsworkspace PUBLIC 
+target_include_directories(artsworkspace PUBLIC
   ${ARTS_SOURCE_DIR}/3rdparty/invlib/src
   ${NETCDF_INCLUDE_DIR}
-  )
+)
 
-add_dependencies (artsworkspace auto_version_cc)
-if (ENABLE_PCH)
+add_dependencies(artsworkspace auto_version_cc)
+
+if(ENABLE_PCH)
   target_precompile_headers(artsworkspace PRIVATE pch_artsworkspace.h)
 endif()
 
 target_link_libraries(artsworkspace PUBLIC auto_partfun xmliobase artscore workspace_internals)
 
-if (NETCDF_FOUND)
-  target_link_libraries (artsworkspace PUBLIC ${NETCDF_LIBRARIES})
-endif (NETCDF_FOUND)
+if(NETCDF_FOUND)
+  target_link_libraries(artsworkspace PUBLIC ${NETCDF_LIBRARIES})
+endif(NETCDF_FOUND)
 
-if (ENABLE_MPI)
-  target_link_libraries (artsworkspace PUBLIC ${MPI_CXX_LIBRARIES})
-endif (ENABLE_MPI)
+if(ENABLE_MPI)
+  target_link_libraries(artsworkspace PUBLIC ${MPI_CXX_LIBRARIES})
+endif(ENABLE_MPI)
 
-if (ENABLE_RT4)
-  target_link_libraries (artsworkspace PUBLIC rt4)
-endif (ENABLE_RT4)
+if(ENABLE_RT4)
+  target_link_libraries(artsworkspace PUBLIC rt4)
+endif(ENABLE_RT4)
 
-if (ENABLE_FASTEM)
-  target_link_libraries (artsworkspace PUBLIC fastem)
-endif (ENABLE_FASTEM)
+if(ENABLE_FASTEM)
+  target_link_libraries(artsworkspace PUBLIC fastem)
+endif(ENABLE_FASTEM)
 
-if (ENABLE_TMATRIX)
-  target_link_libraries (artsworkspace PUBLIC tmatrix)
-endif (ENABLE_TMATRIX)
+if(ENABLE_TMATRIX)
+  target_link_libraries(artsworkspace PUBLIC tmatrix)
+endif(ENABLE_TMATRIX)
 
-########### next target ###############
+# ########## next target ###############
+add_library(xmliobase STATIC
+  xml_io_base.cc
+  xml_io_general_types.cc
+)
 
-add_library (xmliobase STATIC
-        xml_io_base.cc
-        xml_io_general_types.cc
-        )
+target_link_libraries(xmliobase PUBLIC binio matpack)
 
-target_link_libraries (xmliobase PUBLIC binio matpack)
+# #######################################################################################
 
-########################################################################################
-
-########### IPO / LTO support
-if (IPO_SUPPORTED)
+# ########## IPO / LTO support
+if(IPO_SUPPORTED)
   set_property(TARGET artsworkspace PROPERTY INTERPROCEDURAL_OPTIMIZATION TRUE)
   set_property(TARGET binio PROPERTY INTERPROCEDURAL_OPTIMIZATION TRUE)
   set_property(TARGET species PROPERTY INTERPROCEDURAL_OPTIMIZATION TRUE)
