+++ src/workspace_wsv_copy_ctor.cpp
@@ -1,0 +1,3 @@
+#include "auto_wsg.h"
+
+Wsv::Wsv(const Wsv&) = default;
