+++ src/workspace_wsv_copy_fn.cpp
@@ -1,0 +1,7 @@
+#include "auto_wsg.h"
+
+Wsv Wsv::copy() const {
+  return std::visit([](const auto& val) -> Wsv {
+    return Wsv{*val};
+  }, value);
+}
