+++ src/workspace_wsv_copy_set.cpp
@@ -1,0 +1,3 @@
+#include "auto_wsg.h"
+
+Wsv& Wsv::operator=(const Wsv&) = default;
