+++ src/workspace_wsv_ctor.cpp
@@ -1,0 +1,3 @@
+#include "auto_wsg.h"
+
+Wsv::Wsv() : value(std::make_shared<Any>()) {}
