+++ src/workspace_wsv_dtor.cpp
@@ -1,0 +1,3 @@
+#include "auto_wsg.h"
+
+Wsv::~Wsv() = default;
