+++ src/workspace_wsv_move_ctor.cpp
@@ -1,0 +1,3 @@
+#include "auto_wsg.h"
+
+Wsv::Wsv(Wsv&&) noexcept = default;
