+++ src/workspace_wsv_move_set.cpp
@@ -1,0 +1,3 @@
+#include "auto_wsg.h"
+
+Wsv& Wsv::operator=(Wsv&&) noexcept = default;
